requirements.txt
tqdm==4.50.2
numpy==1.19.2
Pillow==8.0.0
